% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{addExprData}
\alias{addExprData}
\title{Function to add expression data to SCE object.}
\usage{
addExprData(sce, exprMat, datName)
}
\arguments{
\item{sce}{Single Cell Experiment Object}

\item{exprMat}{Matrix or object that can be coerced to a matrix containing
Feature count data.}

\item{datName}{String Containing a unique ID for the expression matrix (\code{exprMat})}
}
\value{
The function returns a SCE object with the given expression matrix
stored under the name \code{datName} in the assays slot of the SCE object.
}
\description{
\code{addExprData} is a helper function used to add expression matrices to an
SCE object.
}
\details{
This function accepts a SCE object, an expression matrix, and a name for the
expression matrix. The expression matrix is typically Gene count data stored
in either a monocle or seurat object. This expression matrix is first
converted into a sparse matrix. Then cells and features that do not appear
in the matrix are added with 0 counts as values. Cells and Features may be
missing due to them being present in only one data set (Seurat or Monocle).
Once the missing cell barcodes and features have been added the sparse
expression matrix is then stored in the SCE object under the given name
(\code{datName}).
}
\examples{

\dontrun{

sce <- SingleCellExperiment::SingleCellExperiment()

counts <- matrix(rnorm(100), 10, 10)

sce <- addExprData(sce, counts, "Counts")
}

}
