% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceTools.R
\name{addFeatureExpr}
\alias{addFeatureExpr}
\title{Adds gene expression to plot data.}
\usage{
addFeatureExpr(plotData, feature, redMethod, inputSCE)
}
\arguments{
\item{plotData}{Data Frame which contains the cell embeddings and meta data
for the current plot.}

\item{feature}{String of the gene name chosen by the user. Expression values
for this gene will be added to the plotData data frame.}

\item{redMethod}{String The current dimension reduction method being
visualized. This value is used to determine which gene expression data to use.}

\item{inputSCE}{SingleCellExperiment object which contains data being used.}
}
\value{
The function returns the given data frame with a column added which
contains the expression values for the user chosen gene.
}
\description{
\code{addFeatureExpr} is a function used to add the expression values
for a chosen gene to be added to the data frame which contains all of the
data needed to construct the plot.
}
\examples{

\dontrun{


plotData <- makePlotData(blah, blah, blah)

plotData <- addFeatureExpr(foo, bar, foo)

}

}
