% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{addMissingFeatures}
\alias{addMissingFeatures}
\title{Function to add missing features to a matrix.}
\usage{
addMissingFeatures(dataMat, features)
}
\arguments{
\item{dataMat}{Matrix}

\item{features}{List of all features expected}
}
\value{
Matrix with missing features added
}
\description{
Function to add missing features to a matrix.
}
