% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{addMonocleData}
\alias{addMonocleData}
\title{Add monocle data to a SCE object}
\usage{
addMonocleData(sce, monocleObj)
}
\arguments{
\item{sce}{Single Cell Experiment in which data will be stored}

\item{monocleObj}{Monocle CellDataSet Contains data to transfer to SCE object}
}
\value{
This function returns a SCE object containing the data from the given
monocle object.
}
\description{
\code{addMonocleData} is a function used to transfer data from a monocle
object to a SCE object.
}
\details{
This function is the main function used to transfer data from a given
monocle object into a given SCE object. The expression count data from the
monocle object is added to the assays slot of the SCE object. Currently the
only cell embeddings transferred from the monocle object are the 'Pseudotime'
embeddings from the reducedDimS slot of the monocle object. If other dimension
reduction methods were used they will not be included using this function.
The meta data generated by monocle is added to the colData slot of the SCE
object. Monocle also generates feature level meta data which is currently not
transferred using this function. All names are suffixed with '.Monocle' before
being stored in the SCE object.
}
\examples{

\dontrun{

sce <- SingleCellExperiment()

sceData <- addMonocleData(sce, monocleObj)
}

}
