% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{cloneChoiceUI}
\alias{cloneChoiceUI}
\title{Module which defines clone selection choices for user input.}
\usage{
cloneChoiceUI(id)
}
\arguments{
\item{id}{String which defines the namespace for the module}
}
\value{
Returns a shiny UI element
}
\description{
This module is used to define the UI of the clone selection user input for
CellTagViz. This module is a drop down selection box of the list of clones
present in the data being visualized. Using the selection from this input the
data will be subset in order to only visualize the cells in the selected
clones. This is a simple module which can be used in many situations.
}
