% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{combineSCE}
\alias{combineSCE}
\title{Function used to combine multiple SingleCellExperiment objects.}
\usage{
combineSCE(dataSets)
}
\arguments{
\item{dataSets}{Named List}
}
\value{
SingleCellExperiment
}
\description{
This function takes a named list of data objects from monocle, seuratv2, or
seuratv3. It is important that the list is named as the names are used to
determine how to handle each object.
}
\details{
The function accepts a list of single-cell data objects. This list is then
converted into a list of SingleCellExperiment objects.
}
