% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{contourButtonUI}
\alias{contourButtonUI}
\title{Module which defines a checkbox to add contour lines.}
\usage{
contourButtonUI(id)
}
\arguments{
\item{id}{String which defines the namespace for the module}
}
\description{
This module is used to define the UI of the contour option checkbox for
CellTagViz. This module is a checkbox which add contour lines to the current
plot when ticked. This is a simple module which can be
used for many visualizations.
}
