% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModulesServer.R
\name{createPlot}
\alias{createPlot}
\title{This is a shiny server function for generating CellTagViz plots.}
\usage{
createPlot(input, output, session, plotOptions)
}
\arguments{
\item{input}{Default argument required for shiny modules.}

\item{output}{Default argument required for shiny modules.}

\item{session}{Default argument required for shiny modules.}

\item{plotOptions}{List which contains the user input values.}
}
\value{
Returns plots based on user input values and current tab selection.
}
\description{
\code{createPlot} is a function called in the server function of the
CellTagViz shiny app. The function assigns the reactive user inputs as a
variable. These user inputs are then parsed and passed as arguments to the
appropriate plotting functions. This function takes in the user input, parses
the user input, and passes the input to the correct function depending on
which tab is currently being displayed.
}
