% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{createSCE}
\alias{createSCE}
\title{Switch function used to create SingleCellExperiment Objects}
\usage{
createSCE(scData, objName)
}
\arguments{
\item{scData}{Object containing Single-cell data}

\item{objName}{Unique name for object}
}
\value{
A SingleCellExperiment Object
}
\description{
This function takes in a data object from monocle, seuratv2, or seuratv3 as
input as well as the name of the given object. The function used to convert
the given object into a SingleCellExperiment object is determined using the
given name of the data object.
}
