% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{dataPanelUI}
\alias{dataPanelUI}
\title{Module which defines the Data panel.}
\usage{
dataPanelUI(id)
}
\arguments{
\item{id}{String which defines the namespace for the module}
}
\value{
Returns a shiny UI element
}
\description{
This module is used to define the UI of the data panel for CellTagViz.
This panel will include the option to download the data used by the shiny
app. This is also a good place to include links to thinkgs like GEO or the
protocols.io webpage.
}
