% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{exprAssayChoiceUI}
\alias{exprAssayChoiceUI}
\title{Module which defines a selection input for choosing a dimension reduction
method.}
\usage{
exprAssayChoiceUI(id, inputSCE)
}
\arguments{
\item{id}{String which defines the namespace for the module}

\item{inputSCE}{SingleCellExperiment object which contains current data.}
}
\value{
Returns a shiny UI element
}
\description{
This module is used to define the UI of the dimension reduction selection
input for CellTagViz. This module is a drop down list which allows ths user
to choose which dimension reduction method to visualize. The choices for this
list are the names of the datasets returned by \code{reducedDimNames}.
This is a simple module which can be used for many visualizations.
}
