% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{factorButtonUI}
\alias{factorButtonUI}
\title{Module which defines a checkbox to consider meta data variables as factors.}
\usage{
factorButtonUI(id)
}
\arguments{
\item{id}{String which defines the namespace for the module}
}
\value{
Returns a shiny UI element
}
\description{
This module is used to define the UI of the factor option checkbox for
CellTagViz. This module is a checkbox which when ticked will consider the
given meta data variable as a factor. In turn the variable is considered
categorical and treated as such. This is a simple module which can be
used for many visualizations.
}
