% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{geneChoiceUI}
\alias{geneChoiceUI}
\title{Module which defines feature selection choices for user input.}
\usage{
geneChoiceUI(id)
}
\arguments{
\item{id}{String which defines the namespace for the module}
}
\description{
This module is used to define the UI of the feature selection user input for
CellTagViz. This module is a drop down selection box of the list of features
present in the data being visualized. These features can then be visualized.
This is a simple module which can be used in many situations.
}
