% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{getCellBarcodes}
\alias{getCellBarcodes}
\title{Returns vector of Cell Barcodes.}
\usage{
getCellBarcodes(seurat, monocle)
}
\arguments{
\item{seurat}{Seurat Object}

\item{monocle}{Monocle Object}
}
\value{
THe function returns the union of the cell barcodes present in the
given Monocle and Seurat objects.
}
\description{
\code{getCellBarcodes} is a helper function used to generate a list of Cell
Barcodes from Monocle and Seurat objects.
}
\details{
This function accepts a Seurat and Monocle object, finds the union of the
cell barcodes in each object, and returns the character vector of the union
of the cell barcodes in both the Monocle and Seurat objects.
}
\examples{

\dontrun{

cellBCs <- getCellBarcodes(seuratObj, monocleObj)
}

}
