% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceTools.R
\name{getEmbeddings}
\alias{getEmbeddings}
\title{Access the cell embeddings of SCE object.}
\usage{
getEmbeddings(sce, redMethod, cells = FALSE)
}
\arguments{
\item{sce}{SingleCellExperiment Object}

\item{redMethod}{String Name of dimension reduction method to visualize.
Must be one of methods listed in \code{names(sce@reducedDim)}}

\item{cells}{Character Vector of cells used to subset data. (Optional)}
}
\value{
The function returns the cell embeddings stored in the given slot of
the SingleCellExperiment object.
}
\description{
\code{getEmbeddings} returns the coordinates of cells after dimension
reduction.
}
\details{
This is a function used to subset SCE objects. The output is used to generate
a data frame which is compatible with ggplot. The cell embeddings are then
combined with metadata to generate visualizations with ggplot2.
}
\examples{

sce <- SingleCellExperiment::SingleCellExperiment()

sce@reducedDims@listData$tSNE <- matrix(data = rnorm(100),
                                          nrow = 10,
                                          ncol = 10)

cellEmbeddings <- CellTagViz:::getEmbeddings(sce, "tSNE")

}
