% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{getFeatures}
\alias{getFeatures}
\title{Returns character vector of Feature Names.}
\usage{
getFeatures(seurat, monocle)
}
\arguments{
\item{seurat}{Seurat Object}

\item{monocle}{Monocle Object}
}
\value{
THe function returns the union of the feature names present in the
given Monocle and Seurat objects.
}
\description{
\code{getFeatures} is a helper function used to generate a list of the feature
names from Monocle and Seurat objects.
}
\details{
This function accepts a Seurat and Monocle object, finds the union of the
feature names in each object, and returns the character vector of the union
of the feature names present in both the Monocle and Seurat objects.
}
\examples{

\dontrun{

featureList <- getFeatures(seuratObj, monocleObj)

}

}
