% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceTools.R
\name{getMetaData}
\alias{getMetaData}
\title{Function used to access meta data (colData) of an SCE object.}
\usage{
getMetaData(sce, varName, cells = FALSE)
}
\arguments{
\item{sce}{SingleCellExperiment Object}

\item{varName}{String Name of the variable to return from meta/col data.}

\item{cells}{Character Vector of Cell Barcodes used to subset data.}
}
\value{
This function returns given meta data (\emph{varName}) for the given
cells (\emph{cells}) of a SCE object (\emph{sce}).
}
\description{
\code{getMetaData} returns the meta data variable \emph{varName} of the SCE
object \emph{sce} for the given vector of cells, \emph{cells}.
}
\examples{

\dontrun{

metaData <- sample(letters, 15)

sce <- SingleCellExperiment::SingleCellExperiment(colData = list(
  Letters = metaData
))

letters <- getMetaData(sce, "Letters")
}

}
