% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{initializeSCE}
\alias{initializeSCE}
\title{Initializes SCE object for Monocle and Seurat Data.}
\usage{
initializeSCE(cellBCs, features)
}
\arguments{
\item{cellBCs}{Character Vector of Cell Barcodes. Typically the output from
the function \code{getCellBarcodes}.}

\item{features}{Character Vector of Feature Names. Typically the output
from the function \code{getFeatures}}
}
\value{
The function returns an empty SCE object with the given cell barcodes
and feature names. The SCE object can then be used to add Monocle and Seurat
data to.
}
\description{
\code{initializeSCE} is a helper function used to initialize an SCE object
with the correct cell barcodes and feature list.
}
\details{
This function takes a list of cell barcodes and feature names and creates
an empty SCE object with the given cell barcodes and features. This empty
SCE object is then utilized to combine and add Monocle and Seurat data.
}
\examples{

\dontrun{

cellBCs <- getCellBarcodes(seuratObj, monocleObj)

featureList <- getFeatures(seuratObj, monocleObj)

sce <- initializeSCE(cellBCs, featureList)
}

}
