% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{makeDataMatrix}
\alias{makeDataMatrix}
\title{Creates sparse matrices of feature count data}
\usage{
makeDataMatrix(data2add, cellBCs, features)
}
\arguments{
\item{data2add}{Matrix-like object of feature counts}

\item{cellBCs}{Vector of all Cell Barcodes in the SCE}

\item{features}{Vector of all features in the SCE}
}
\value{
The function returns a sparse matrix of feature counts to be stored
in a SCE object
}
\description{
\code{makeDataMatrix} accepts a matrix like object of feature count data
and returns a sparse matrix of the feature counts.
}
\details{
This is a helper function for adding feature counts to SCE objects.
A matrix of feature counts typically from a Seurat or Monocle object is
accepted as an argument. This matrix is then coerced into a sparse matrix and
missing cell barcodes and/or features are identified. The missing cell
barcodes and features are then added to the matrix with 0 values. The
completed matrix is then returned and can be stored as data in the assays
slot of an SCE object.
}
\examples{

barcodes <- c("BC1", "BC2")

genes <- c("Apoa1", "Mettl7a1")

exprMat <- matrix(rnorm(100), 3, 3)

newMat <- makeDataMatrix(exprMat, barcodes, genes)

}
