% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{makeSCEMonocle}
\alias{makeSCEMonocle}
\title{Function to create SingleCellExperiment object using a monocle CDS.}
\usage{
makeSCEMonocle(monocleObj)
}
\arguments{
\item{monocleObj}{CellDataSet Object with single-cell Data}
}
\value{
A SingleCellExperiment object
}
\description{
This function takes a monocle CDS object as input and converts the
CellDataSet object into a SingleCellExperiment object.
}
