% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{makeSCESeuratV3}
\alias{makeSCESeuratV3}
\title{Function to create SingleCellExperiment Object.}
\usage{
makeSCESeuratV3(seuratObj)
}
\arguments{
\item{seuratObj}{SeuratV3 object}
}
\value{
A SingleCellExperiment Object
}
\description{
This function takes a Seurat V3 object as input and creates a
SingleCellExperiment Object using the data stored in the seurat object.
}
