% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{makeVizData}
\alias{makeVizData}
\title{Creates CellTagViz compatable data sets.}
\usage{
makeVizData(dataSets)
}
\arguments{
\item{dataSets}{List This is a named list which contains the Seurat and Monocle
objects. The names for the list must be Seurat or Monocle corresponding to the
object.}
}
\value{
The function returns a SCE object with the data from Monocle and Seurat
objects combined.
}
\description{
\code{makeVizData} currently accepts Seurat and Monocle objects and returns
a SingleCellExperiment object for use with CellTagViz.
}
\details{
This function accepts Seurat and Monocle objects. The data from these objects
is then combined and used to create a SingleCellExperiment object. This object
is compatable with CellTagViz and allows simple visualiztion of the analysis
performed with Seurat and Monocle.
}
\examples{

\dontrun{

dataSets <- list(seurat = seuratObj, monocle = monocleObj)

sce <- makeVizData(dataSets)

}

}
