% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{metaChoiceUI}
\alias{metaChoiceUI}
\title{Module which defines meta data variable  selection choices for user input.}
\usage{
metaChoiceUI(id, inputSCE)
}
\arguments{
\item{id}{String which defines the namespace for the module}

\item{inputSCE}{SingleCellExperiment object which contains current data.}
}
\value{
Returns a shiny UI element
}
\description{
This module is used to define the UI of the meta data selection user input
for CellTagViz. This module is a drop down selection box of the list of
meta data variables which can be visualized. This is a simple module which
can be used in many situations.
}
