% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plotDownloadUI}
\alias{plotDownloadUI}
\title{Module which defines a button to download the current plot as a PDF.}
\usage{
plotDownloadUI(id, inputSCE)
}
\arguments{
\item{id}{String which defines the namespace for the module}

\item{inputSCE}{SingleCellExperiment object which contains current data.}
}
\value{
Returns a shiny UI element
}
\description{
This module is used to define the UI of the plot download button for
CellTagViz. This module is a button which when clicked will download and save
the current plot for the user. Slight differences in the downloaded plot and
the plot produced by CellTagViz might occur. This is due to the CellTagViz
interactive plot being rendered via plotly while the downloaded plot is
rendered with ggplot2. This is a simple module which can be used to download
many visualizations.
}
