% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plotPanelSideBar}
\alias{plotPanelSideBar}
\title{Module which defines the sidebar of the Plot Panel.}
\usage{
plotPanelSideBar(id, inputSCE)
}
\arguments{
\item{id}{String which defines the namespace for the module}

\item{inputSCE}{SingleCellExperiment object which contains current data.}
}
\value{
Returns a shiny UI element
}
\description{
This module is used to define the UI of the Plot Panel sidebar for
CellTagViz. This module is a taglist which contains calls to multiple other
modules. The function calls this module makes define conditional sidebar
layouts for each of the tabs in the tabset on the Plot Panel. By defining
multiple sidebar layouts with the function calls allows each tab to have a
unique sidebar. This module also calls the function \code{plotDownloadUI}
this was done in this module as I could not produce proper formatting when
placing the module call inside the modules which define the conditional
sidebars. Furthermore, because the option to download plots is available for
each of the plots the download button is not neccesarrily conditional. This
makes it possible to call the module \code{plotDownload} in this module and
generate the desired layout. Sidebars for new plot tabs can be easily added
by creating a new module defining the sidebar and placing the function call
in this module.
}
