% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModulesServer.R
\name{plotTsne}
\alias{plotTsne}
\title{Parses shiny user input and plots dimension reduction visualiztions.}
\usage{
plotTsne(sce, feature = FALSE, metaVar = FALSE, clones = FALSE,
  factor = FALSE, contour = FALSE, redMethod = FALSE,
  exprDat = FALSE)
}
\arguments{
\item{sce}{Yes}

\item{feature}{Yes}

\item{metaVar}{Yes}

\item{clones}{Yes}

\item{factor}{Yes}

\item{contour}{Yes}

\item{redMethod}{Yes}

\item{exprDat}{Yes}
}
\value{
This function returns a ggplot2 object
}
\description{
This function generates 2d visualizations of single-cell RNA data. The
function parses the user input and plots the visualization according to the
given user inputs.
}
