% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{scatterSideBarUI}
\alias{scatterSideBarUI}
\title{This module defines the UI element for the Scatter Chart panel Sidebar.}
\usage{
scatterSideBarUI(id)
}
\arguments{
\item{id}{String which defines the namespace for the module}
}
\value{
Returns a Shiny UI element
}
\description{
This module defines a conditional panel which is displayed when the Scatter
chart plot panel is displayed. It makes calls to modules which allow the
user to choose a meta data variable to color the plot, choose a feature to
visualize the expression, and ensure categorical variables are treated as
such. The module also defines help text which is rendered in the panel. This
is used to ensure the correct sidebars are displayed on the correct panels
and is only used during testing. Importantly this panel is set to display
only in the correct circumstances. This sidebar will only be displayed when
\code{input$plotPanel} is equal to "Scatter Plots". This module can be
easily modified to define new sidebar panels for new plots.
}
